Optimizer speed improvements
If one piece of gear is overall worse than another in the same slot, we filter that out before we start doing the brute force through combinations.

This gives a correctness improvement as well:  As shown in some of the tests, there are small saddle points with max strength vs speed that the old optimizer couldn't see the best solution to.  We're still using that optimizer, and there's still the potential for issues, but a lot of the bad possibilities are gone so slightly more DPS is visible in certain situations.