Fix `useStore` with an even number of pending updates
This changed some time between preact 10.6.x and preact 10.10.x, to match the latest React.  Using a bool-flipping forceUpdate can lose a render if it's called more than once, but an even number of times, between renders.  In those cases, the new value is === to the old one and it bails out.