Fix bug with URL decoding and second character only.
Second job only didn't work; you got the job, but no licenses.  So for instance, ?AQAAgAAgAAhAAAgQZPEQEHBAAg makes Vaan a white mage, and gives him Quickening 1-3.  In earlier versions, it makes him a white mage, but gives him nothing else.  The encoder never emitted second-job-only URLs, and this is only a bug because the Urls.md documentation claims it should work.